+++ src/shell/lib/window_manager.c
@@ -5,8 +5,8 @@
 
 // Note: Would be interesting to make this a dynamically linked library 
 void window_init(window_info* winfo){
-    winfo->main_buffer = (uint8_t*) malloc (sizeof(uint8_t) * ((uint32_t)winfo->xlen) * ((uint32_t)winfo->ylen));
-    winfo->rear_buffer = (uint8_t*) malloc (sizeof(uint8_t) * ((uint32_t)winfo->xlen) * ((uint32_t)winfo->ylen));
+    winfo->main_buffer = (uint32_t*) malloc (sizeof(uint32_t) * ((uint32_t)winfo->xlen) * ((uint32_t)winfo->ylen));
+    winfo->rear_buffer = (uint32_t*) malloc (sizeof(uint32_t) * ((uint32_t)winfo->xlen) * ((uint32_t)winfo->ylen));
 }
 
 void window_clear(window_info* winfo){
@@ -14,7 +14,7 @@ void window_clear(window_info* winfo){
     free(winfo->rear_buffer);
 }
 
-void window_draw_pixel(window_info* winfo, uint16_t row, uint16_t col, uint8_t color){
+void window_draw_pixel(window_info* winfo, uint16_t row, uint16_t col, uint32_t color){
     winfo->main_buffer[row * winfo->xlen + col] = color;
 }
